// src/app/models/Character.ts
export class Character {
    name:       string;
    height:     string;
    mass:       string;
    hair_color: string;
    skin_color: string;
    eye_color:  string;
    birth_year: string;
    gender:     string;
    homeworld:  any;
    films:      any[];
    species:    any[];
    vehicles:   string[];
    starships:  any[];
    created:    Date;
    edited:     Date;
    url:        string;
}